third_party/WebKit/LayoutTests/fast/dom/DeviceMotion/create-event-expected.txt

Tests that document.createEvent() works with DeviceMotionEvent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof event == 'object' is true
PASS event.__proto__ is DeviceMotionEvent.prototype
PASS 'type' in event is true
PASS 'bubbles' in event is true
PASS 'cancelable' in event is true
PASS 'acceleration' in event is true
PASS 'accelerationIncludingGravity' in event is true
PASS 'rotationRate' in event is true
PASS 'interval' in event is true
PASS typeof newEvent.type == 'string' is true
PASS newEvent.type is "devicemotion"
PASS typeof newEvent.bubbles == 'boolean' is true
PASS event.bubbles is false
PASS newEvent.bubbles is false
PASS typeof newEvent.cancelable == 'boolean' is true
PASS event.cancelable is false
PASS newEvent.cancelable is false
PASS typeof event.acceleration == 'object' is true
PASS typeof event.accelerationIncludingGravity == 'object' is true
PASS typeof event.rotationRate == 'object' is true
PASS typeof event.interval == 'object' is true
PASS successfullyParsed is true

TEST COMPLETE
